Fix: Schedule tasks using Celery instead of web2py.
