--- models/scheduler.py
@@ -1,8 +1,0 @@
-from gluon.scheduler import Scheduler
-from gluon import current
-
-from feedback import _scheduled_builder  # noqa: F401
-
-if settings.academy_mode:
-    scheduler = Scheduler(db, migrate=table_migrate_prefix, heartbeat=1)
-    current.scheduler = scheduler
