+++ requirements.txt
@@ -1,3 +1,4 @@
+celery[redis,gevent]
 diff-match-patch>=20110725.1
 lxml>=3.3.5
 numpy>=1.9.2
